int64 warnings

+++ src/util/recoder/recode_dll.h
@@ -30,30 +30,30 @@ enum ConvEncodingEnum{
   CONV_ENCODING_ISO8895_5 = 4
 };
 
-RECODE_DLL_API int RECODE_DECL ConvertUCS2ToMultibyte(const short* ucs2, int ucs2buff_size,
-                                          char* text, int textbuff_size,
+RECODE_DLL_API int RECODE_DECL ConvertUCS2ToMultibyte(const short* ucs2, size_t ucs2buff_size,
+                                          char* text, size_t textbuff_size,
                                           ConvEncodingEnum encoding);
-RECODE_DLL_API int RECODE_DECL ConvertMultibyteToUCS2(const char* text, int textbuff_size,
-                                          short* ucs2, int ucs2buff_size,
+RECODE_DLL_API int RECODE_DECL ConvertMultibyteToUCS2(const char* text, size_t textbuff_size,
+                                          short* ucs2, size_t ucs2buff_size,
                                           ConvEncodingEnum encoding);
 
-RECODE_DLL_API int RECODE_DECL Convert7BitToText(const char* bit7buf, int bit7buf_size,
-                                     char* text, int textbuff_size);
+RECODE_DLL_API int RECODE_DECL Convert7BitToText(const char* bit7buf, size_t bit7buf_size,
+                                     char* text, size_t textbuff_size);
 
-RECODE_DLL_API int RECODE_DECL ConvertTextTo7Bit(const char* text, int textbuff_size,
-                                     char* bit7buf, int bit7buf_size,
+RECODE_DLL_API int RECODE_DECL ConvertTextTo7Bit(const char* text, size_t textbuff_size,
+                                     char* bit7buf, size_t bit7buf_size,
                                      ConvEncodingEnum encoding);
-RECODE_DLL_API int RECODE_DECL Convert7BitToUCS2(const char* bit7buf, int bit7buf_size,
-                                     short* ucs2, int ucs2buff_size);
+RECODE_DLL_API int RECODE_DECL Convert7BitToUCS2(const char* bit7buf, size_t bit7buf_size,
+                                     short* ucs2, size_t ucs2buff_size);
 
-RECODE_DLL_API int RECODE_DECL ConvertUCS2To7Bit(const short* ucs2, int ucs2buff_size,
-                                     char* bit7buf, int bit7buf_size);
+RECODE_DLL_API int RECODE_DECL ConvertUCS2To7Bit(const short* ucs2, size_t ucs2buff_size,
+                                     char* bit7buf, size_t bit7buf_size);
 
-RECODE_DLL_API int RECODE_DECL Transliterate(const char* buf,int len,
-                    ConvEncodingEnum encoding,char *dest,int destlen);
+RECODE_DLL_API int RECODE_DECL Transliterate(const char* buf,size_t len,
+                    ConvEncodingEnum encoding,char *dest,size_t destlen);
 
-RECODE_DLL_API unsigned RECODE_DECL ConvertLatin1ToSMSC7Bit(const char* in, unsigned chars,char* out);
-RECODE_DLL_API unsigned RECODE_DECL ConvertSMSC7BitToLatin1(const char* in, unsigned chars,char* out);
+RECODE_DLL_API unsigned RECODE_DECL ConvertLatin1ToSMSC7Bit(const char* in, size_t chars,char* out);
+RECODE_DLL_API unsigned RECODE_DECL ConvertSMSC7BitToLatin1(const char* in, size_t chars,char* out);
 
 
 }
